Move iterateObserversSafely into separate module.
